update: Added the code for the finance_cal file

+++ python_introduction/finance_calculator.py
@@ -2,8 +2,8 @@ monthly_income = input('Enter your monthly income: ')
 total_monthly_expenses = input('Enter your total monthly expenses: ')
 annual_interest_rate = 0.05
 
-monthly_income = int(monthly_income)
-total_monthly_expenses = int(total_monthly_expenses)
+monthly_income = float(monthly_income)
+total_monthly_expenses = float(total_monthly_expenses)
 
 monthly_savings = float(monthly_income - total_monthly_expenses)
